[Feat]: removed ejs
