--- src/views/inviteMail.ejs
@@ -1,11 +1,0 @@
-<h2>Hey <%= email %></h2>
-You are invited to join <%= organisation %> as
-<% if(manager === 1){ %>
-  <strong>a manager</strong>
-<% } else{ %>
-  <strong>an employee</strong>
-<% } %>
-<br />
-
-Click <a href="<%= URL %>/invite/register?token=<%= token %>"</a>here</a> to register.
-
